Use mmap buffer for playback with option '-M'
If specify the option '-M' for playback, the buffer has been shared from
kernel to user space with mmap(); but when write the audio data it still
uses the ioctl method to transfer data from user space to kernel.

This patch is to check if '-M' flag has been set, then it will directly
write data to mmaped buffer.

Signed-off-by: Leo Yan <leo.yan@linaro.org>